set triger product histories

+++ database/migrations/2022_05_09_120655_create_cities_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('cities', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('expedition_cost');
+            $table->unsignedBigInteger('expedition_cost');
             $table->timestamps();
         });
     }
